# Conflicts:
#	src/assets/styles/main.scss
#	src/components/search/SearchModal.js
#	src/pages/GraphView.js
#	src/pages/Index.js
#	src/pages/profile/GraphTemplates.js
#	src/store/actions/app.js

+++ src/components/AutoSave.js
@@ -5,7 +5,11 @@ import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import Chart from '../Chart';
-import { getSingleGraphRequest, updateGraphPositionsRequest, updateGraphThumbnailRequest } from '../store/actions/graphs';
+import { getSingleGraphRequest, 
+  updateGraphPositionsRequest, 
+  updateGraphThumbnailRequest, 
+  getGraphsListRequest,
+} from '../store/actions/graphs';
 import ChartUtils from '../helpers/ChartUtils';
 import {
   createNodesRequest,
@@ -14,6 +18,7 @@ import {
   updateNodesRequest,
 } from '../store/actions/nodes';
 import { createLinksRequest, deleteLinksRequest, updateLinksRequest } from '../store/actions/links';
+import { toggleDeleteState } from '../store/actions/app';
 import {
   createLabelsRequest,
   deleteLabelsRequest, toggleFolderRequest,
@@ -27,6 +32,7 @@ class AutoSave extends Component {
     match: PropTypes.object.isRequired,
     history: PropTypes.object.isRequired,
     defaultImage: PropTypes.bool.isRequired,
+    deleteState: PropTypes.bool.isRequired,
 
     createNodesRequest: PropTypes.func.isRequired,
     deleteNodesRequest: PropTypes.func.isRequired,
@@ -47,8 +53,10 @@ class AutoSave extends Component {
     updateNodesCustomFieldsRequest: PropTypes.func.isRequired,
     
     getSingleGraphRequest: PropTypes.func.isRequired,
+    getGraphsListRequest: PropTypes.func.isRequired,
 
     updateGraphThumbnailRequest: PropTypes.func.isRequired,
+    getGraphsListRequest: PropTypes.func.isRequired,
   }
 
   async componentDidMount() {
@@ -73,7 +81,9 @@ class AutoSave extends Component {
 
   componentWillUnmount() {
     clearTimeout(this.thumbnailTimeout);
-    this.thumbnailListener();
+    if (typeof(this.thumbnailListener) === "function") {
+      this.thumbnailListener();
+    }
     window.removeEventListener('beforeunload', this.handleUnload);
   }
 
@@ -99,7 +109,7 @@ class AutoSave extends Component {
     if (!Chart.autoSave) {
       return;
     }
-    if (Chart.isLoading === true) {
+    if (Chart.isLoading() === true) {
       return
     }
     this.timeout = setTimeout(this.saveGraph, 0);
@@ -182,6 +192,7 @@ class AutoSave extends Component {
   }
 
   saveGraph = async () => {
+    const deleteState = this.props.deleteState
     const { match: { params: { graphId } } } = this.props;
     if (!graphId || Chart.isAutoPosition) {
       return;
@@ -301,7 +312,7 @@ class AutoSave extends Component {
     if (updateNodes.length) {
       promise.push(this.props.updateNodesRequest(graphId, updateNodes));
     }
-    if (deleteNodes.length) {
+    if (deleteNodes.length && deleteState) {
       promise.push(this.props.deleteNodesRequest(graphId, deleteNodes));
     }
     // if (updateNodePositions.length) {
@@ -323,7 +334,7 @@ class AutoSave extends Component {
     if (updateLinks.length) {
       promise.push(this.props.updateLinksRequest(graphId, updateLinks));
     }
-    if (deleteLinks.length) {
+    if (deleteLinks.length && deleteState) {
       promise.push(this.props.deleteLinksRequest(graphId, deleteLinks));
     }
 
@@ -337,7 +348,7 @@ class AutoSave extends Component {
     // if (updateLabelPositions.length) {
     //   promise.push(this.props.updateLabelPositionsRequest(graphId, updateLabelPositions));
     // }
-    if (deleteLabels.length) {
+    if (deleteLabels.length && deleteState) {
       promise.push(this.props.deleteLabelsRequest(graphId, deleteLabels));
     }
     Chart.event.emit('auto-save');
@@ -354,6 +365,7 @@ class AutoSave extends Component {
        
     });
     document.body.classList.remove('autoSave');
+    this.props.toggleDeleteState(false)
   }
 
   handleUnload = (ev) => {
@@ -363,19 +375,26 @@ class AutoSave extends Component {
   }
 
   handleRouteChange = (newLocation) => {
+    if (Chart.isLoading()) {
+      return
+    }
     const { location } = this.props;
     if (location.pathname !== newLocation.pathname) {
+      Chart.loading(true)
       this.updateThumbnail();
     }
   }
 
   updateThumbnail = async () => {
     const { defaultImage } = this.props
+    const page = 1
+    const order = 'newest'
     document.body.classList.add('autoSave');
     const svg = ChartUtils.getChartSvg();
     const { match: { params: { graphId } } } = this.props;
     if (!defaultImage) {
       await this.props.updateGraphThumbnailRequest(graphId, svg, 'small');
+      this.props.getGraphsListRequest(page, {filter: order})
     }
     document.body.classList.remove('autoSave');
   }
@@ -386,7 +405,8 @@ class AutoSave extends Component {
 }
 
 const mapStateToProps = (state) => ({
-  defaultImage: state.graphs.singleGraph.defaultImage
+  defaultImage: state.graphs.singleGraph.defaultImage,
+  deleteState: state.app.deleteState,
 });
 
 const mapDispatchToProps = {
@@ -410,6 +430,8 @@ const mapDispatchToProps = {
   deleteLabelsRequest,
   toggleFolderRequest,
   getSingleGraphRequest,
+  getGraphsListRequest,
+  toggleDeleteState
 };
 
 const Container = connect(
